Group_4_Submission/requirements.txt
langchain-community
langchain-core
langchain-openai
langgraph
numpy
python-dotenv
scikit-learn
sentence-transformers
streamlit
tavily-python
watchdog
tiktoken
requests
wikipedia
arxiv
beautifulsoup4